+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 import About from './pages/about'
-import {BrowserRouter as Router, Route, Link, Routes} from 'react-router-dom';
-import Gallery from './Galery'
+import {BrowserRouter as Router, Route, Routes} from 'react-router-dom';
+import Gallery from '../src/pages/galeria/Galery'
 import Main from './Components/Corpo/Main';
 import NavBoot from './Components/Navbar/NavBoot'
