+++ src/Components/Corpo/PageTitle/style.css
@@ -1,7 +1,7 @@
 .page-title{
-    margin-left: 40rem;
     font-size: 2.2rem;
     font-family: Magic Retro;
+    margin: 1rem 0;
 }
 
 u{
