--- src/Galery.css
@@ -1,12 +1,0 @@
-.body-galery {
-    margin-top: 20px;
-    display: flex;
-    flex-direction: row;
-}
-
-.draws {
-    display: grid;
-    grid-template-columns: repeat(5, 1fr);
-    gap: 20px;
-    margin-left: 50px;
-}
