+++ src/pages/Sidebar/index.js
@@ -3,10 +3,10 @@ import "./style.css";
 function Sidebar() {
     return (
         <div className="sidebar">
-            <a href="">2022</a>
-            <a href="">2021</a>
-            <a href="">2020</a>
-            <a href="">2019</a>
+            <a href="/">2022</a>
+            <a href="/">2021</a>
+            <a href="/">2020</a>
+            <a href="/">2019</a>
         </div>
     )
 };
