+++ src/pages/Sidebar/style.css
@@ -20,3 +20,17 @@
     background-color: var(--green-color);
     color: white;
 }
+
+@media screen and (max-width:1023px) {
+    .sidebar{
+        border-style: solid;
+        border-right-color: unset;
+        border-bottom-color: var(--green-color);
+        align-items: center;
+        display: flex;
+        flex-direction: row;
+        justify-content: space-evenly;
+        margin: 0;
+        width: 100%;
+    }
+}
