+++ src/pages/about/style.css
@@ -1,26 +1,17 @@
-.imagem {
-    position: absolute;
-    width: 467px;
-    height: 477.08px;
-    left: 166px;
-    top: 398px;
+.texto {
+    color: #1f1e1e;
+    font-family: 'JetBrains Mono';
+    text-align: justify;
 }
 
-.texto {
-    position: absolute;
-    width: 50rem;
-    height: 40rem;
-    left: 50rem;
-    top: 25rem;
+.content{
+    align-items: center;
+    display: flex;
+    justify-content: space-evenly;
+}
 
-    font-family: 'JetBrains Mono';
-    font-style: normal;
-    font-weight: 400;
-    font-size: 35px;
-    line-height: 46px;
-    text-align: right;
-
-    color: #000000;
-
-    text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+@media screen and (max-width:1023px) {
+    .imagem{
+        display: none;
+    }
 }
